+++ app/src/main/java/com/soundcloud/android/model/TrackStats.java
@@ -4,24 +4,24 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class TrackStats extends PlayableStats {
 
-    private long playbackCount;
-    private long commentsCount;
+    private int playbackCount;
+    private int commentsCount;
 
-    public long getPlaybackCount() {
+    public int getPlaybackCount() {
         return playbackCount;
     }
 
     @JsonProperty("playback_count")
-    public void setPlaybackCount(long playbackCount) {
+    public void setPlaybackCount(int playbackCount) {
         this.playbackCount = playbackCount;
     }
 
-    public long getCommentsCount() {
+    public int getCommentsCount() {
         return commentsCount;
     }
 
     @JsonProperty("comments_count")
-    public void setCommentsCount(long commentsCount) {
+    public void setCommentsCount(int commentsCount) {
         this.commentsCount = commentsCount;
     }
 }
